[Refactor] Clean code
